+++ python/pathway/internals/yaml_loader.py
@@ -146,7 +146,17 @@ class Resolver:
         if all(c.upper() or c == "_" for c in v.name):
             s = os.environ.get(v.name)
             if s is not None:
-                return load_yaml(s)
+                # using yaml.Loader instead of PathwayYamlLoader to prevent recursive
+                # parsing of environment variables
+                parsed_value = yaml.load(s, yaml.Loader)
+                if (
+                    isinstance(parsed_value, int)
+                    or isinstance(parsed_value, float)
+                    or isinstance(parsed_value, bool)
+                ):
+                    return parsed_value
+                else:
+                    return s
 
         raise KeyError(f"variable {v} is not defined")